feat(websocket): crate wrappers
