+++ server/src/common/utils/socketsUtils.js
@@ -1,12 +1,8 @@
-const { Server } = require("socket.io");
 const passport = require("passport");
 const cookieParser = require("cookie-parser");
 const bodyParser = require("body-parser");
-const corsRules = require("../../common/corsRules");
 const logger = require("../../common/logger");
-const authMiddleware = require("../middlewares/authMiddleware");
-
-const dossier = require("./dossier");
+const authMiddleware = require("../../http/middlewares/authMiddleware");
 
 const wrapMiddleware = (middleware) => (socket, next) => middleware(socket.request, {}, next);
 
@@ -20,7 +16,7 @@ const ioUseNamespace = (io, name, components, callback) => {
 
   io.of(name).on("connection", (socket) => {
     socket.on("error", (err) => {
-      logger["error"](err, `Socket Request KO`);
+      logger["error"](err, `Socket Request ${name} KO`);
       socket.disconnect();
     });
 
@@ -28,13 +24,6 @@ const ioUseNamespace = (io, name, components, callback) => {
   });
 };
 
-module.exports = async (httpServer, components) => {
-  const io = new Server(httpServer, {
-    cors: {
-      origin: corsRules.origin,
-      allowedHeaders: corsRules.allowedHeaders,
-    },
-  });
-
-  ioUseNamespace(io, "/dossier", components, dossier);
+module.exports = {
+  ioUseNamespace,
 };
